requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9; python_version < "3.14"
psycopg[binary]>=3.2.0; python_version >= "3.14"
fpdf2==2.7.8
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
